chore(desktop): extract engine binary spawn into helper
engine-bridge.ts was at 598/600 before this branch and the
ImplementationPhase wiring in the next commit would push it over.
Split the binary-discovery + child-spawn logic into a new sibling
module (engine-bridge-spawn.ts, ~95 lines) so the bridge file
stays at 545 lines and has room for future additions.

The spawned helper has no dependencies on the bridge's connection
state — it's a one-shot 'find the binary and spawn it' function
that runs at most once per cold start. Moving it out of the
EngineBridge class is a clean seam because the class doesn't read
any of the spawned helper's locals after the spawn returns.

Behavior preserved: same discovery order (bundled .app → dev
monorepo → installed CLI → PATH), same macOS TCC inheritance
strategy, same env-var contract (ION_SOCKET_PATH, ION_PID_PATH).
The full rationale comments from the original _startServer body
moved verbatim to the new file.
